Rename package to testsuites, removed unused file.
